requirements.txt
SQLAlchemy~=1.4.4
requests~=2.25.1
bs4~=0.0.1
beautifulsoup4~=4.9.3
Scrapy==2.5.0
python-dotenv==0.17.0